use-case-3-domain-specific-model-creation/domain-specific-model-pipeline/requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro==0.19.8
kedro-datasets==4.1.0
kedro-viz==9.2.0
kedro[jupyter]
notebook
pytest~=7.2
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
ruff~=0.1.8
scikit-learn==1.5.1
pandas==2.2.2
pyarrow==17.0.0

#Fine-tuning requirements 
bitsandbytes==0.43.3
transformers==4.44.2
peft==0.12.0
accelerate==0.33.0
datasets==2.21.0
trl==0.10.1

#login to hf or access via pvt token
huggingface_hub

# .env file loading
python-dotenv==1.0.1